[MIG]stock_push_delay v13
