+++ stock_push_delay/__manifest__.py
@@ -2,11 +2,12 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 {
     "name": "Stock Push Delay",
-    "summary": "Manual evalaution of Push rules",
-    "version": "11.0.1.0.1",
+    "summary": "Manual evaluation of Push rules",
+    "version": "13.0.1.0.1",
     "category": "Inventory",
     "author": "ForgeFlow, " "Odoo Community Association (OCA)",
     "license": "AGPL-3",
     "website": "https://github.com/OCA/stock-logistics-workflow",
-    "depends": ["stock",],
+    "depends": ["stock"],
+    "development_status": "Alpha",
 }
