+++ stock_push_delay/models/__init__.py
@@ -1,1 +1,2 @@
-from . import stock
+from . import stock_move
+from . import stock_picking
